requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
openai==1.43.0
python-dotenv==1.0.1
